feat(keep-state): add a strategy to keep state between PHP processes
This PR aims to resolve the issue that PHP shares no state between PHP processes.
If the primary DNS server is down, then every new PHP process would have to hit the timeout connection
once before falling back to the other DNS provider.
For now users have to manually state that they want to use the file based caching strategy.
After the robustness of this has been proven, we will make this the default strategy.
Note that if you are working in a RO filesystem, it would automatically fallback to in memory.