Describe actual behavior of scoped copy operation
The comments imply that records will be copied if you specify a scope.
This is not the behavior I observe, and this is also not what is
documented on https://www.algolia.com/doc/rest-api/search/#copymove-index :

> Keep in mind that when you use the scopes parameter, you will no
longer be copying records (objects) but only the specified scopes.

It might have been true in the past, but it is currently not the case.