🛑🎉 Removed Client and Index Interfaces
Next time we need to add a method, which should happen regularily,
we won't be able to update the Interface without breaking compatibilty.
We'll end up with either new major very often or out-of-sync interface.

If you want to modify the client, add or your methods, simply extends
the client or the index. The classes are not final, the signature of
the methods won't change until next major.
will be maintained.