fix: replace expired certificate within embedded certificate chain
We previously embedded a full certificate chain for the sole purpose of
supporting environments where the system certificates are not available
or outdated.

However, on May 30th, 2020, one of the certificates within the
certificate chain we provided expired. This has the consequence of
making OpenSSL <1.1.1 and LibreSSL <3.2.0 consider the legit Algolia
certificate invalid, when it was not.

This commit replaces the faulty certificate from the embedded
certificate chain so that users who depends on it with outdated versions
of OpenSSL or LibreSSL can still prove Algolia servers identity.